Create a caching option in the config
Allows a user to provide the ability to specify a StorageInterface class in the config for a caching system. This will cache GET requests for less server load. In the future, a cache invalidation will be required upon PUT/POST/DELETE for the URI in question. I am choosing to specify the cache in the config as opposed to a parameter in the Hybridauth object because I feel that this is more of a global setting to the application than an instance-level setting. That being said, I am not opposed to treating the cache as a parameter in the constructor.